movies/notification/requirements.txt
fastapi==0.110.0
orjson==3.9.15
pydantic==2.5.2
uvicorn==0.28.0
gunicorn==21.2.0
pydantic-settings==2.2.1
#async-fastapi-jwt-auth==0.6.4
pymongo==4.6.2
motor==3.3.2

 opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
